+++ test/app/components/empty_component.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class EmptyComponent < ViewComponent::Base
+  def call; end
+end